+++ TCAPP.Domain/RelationalData/ContentTaxonomy.cs
@@ -5,8 +5,8 @@ namespace TCAPP.Domain.RelationalData
 {
     public class ContentTaxonomy
     {
-        public decimal IdContent { get; set; }
-        public decimal IdTaxonomy { get; set; }
+        public long IdContent { get; set; }
+        public long IdTaxonomy { get; set; }
         public Content Content { get; set; }
         public Taxonomy Taxonomy { get; set; }
     }
