+++ TCAPP.Domain/RelationalData/ParentContent.cs
@@ -4,8 +4,8 @@ namespace TCAPP.Domain.RelationalData
 {
     public class ParentContent
     {
-        public decimal IdContent { get; set; }
-        public decimal IdParentContent { get; set; }
+        public long IdContent { get; set; }
+        public long IdParentContent { get; set; }
         public Content Content { get; set; }
         public Content Parent { get; set; }
     }
