+++ TCAPP.Domain/RelationalData/ParentTaxonomy.cs
@@ -4,8 +4,8 @@ namespace TCAPP.Domain.RelationalData
 {
     public class ParentTaxonomy
     {
-        public decimal IdTaxonomy { get; set; }
-        public decimal IdParentTaxonomy { get; set; }
+        public long IdTaxonomy { get; set; }
+        public long IdParentTaxonomy { get; set; }
         public Taxonomy Taxonomy { get; set; }
         public Taxonomy Parent { get; set; }
     }
