+++ TCAPP.Domain/RelationalData/UserContent.cs
@@ -8,10 +8,10 @@ namespace TCAPP.Domain.RelationalData
 {
     public class UserContent
     {
-        public decimal IdUser { get; set; }
-        public decimal IdContent { get; set; }
-        public decimal IdContentRelationType { get; set; }
-        public decimal? IdCollection { get; set; }
+        public long IdUser { get; set; }
+        public long IdContent { get; set; }
+        public long IdContentRelationType { get; set; }
+        public long? IdCollection { get; set; }
         public User User { get; set; }
         public Content Content { get; set; }
         public ContentRelationType ContentRelationType { get; set; }
